+++ index-page/src/components/reporter-intro.js
@@ -157,7 +157,7 @@ class ReporterIntro extends React.PureComponent {
 
   render() {
     const authorHref = 'authors'
-    const donationHref = externalLinks.monthlyDonation
+    const donationHref = externalLinks.donation
     const hiringHref = 'a/hiring-job-description'
     return (
       <ContentContainer